Meeting on 2019-02-03
Justin reported the bucket was not going up and down.  Reviewed the code and it appears that the 2019-01-30 check-in accidentally deleted several meetings' worth of code.

We copied the items that pertained to the intake motor from the 2019-01-30 code, reverted to the 2019-01-08 code and then put those intake motor items back into the code and tested.

Doing interim check-in to make future evaluations less confusing.